+++ spec/service_operator/step_spec.rb
@@ -1,7 +1,12 @@
 # frozen_string_literal: true
 
 RSpec.describe ServiceOperator::Step do
-  let(:operator) { Class.new.send(:include, ServiceOperator) }
+  let(:operator) { Class.new.send(:include, ServiceOperator::Helpers) }
+  let(:configuration) { ServiceOperator::Configuration.new }
+
+  before do
+    allow(operator).to receive(:configuration).and_return(configuration)
+  end
 
   describe '.run' do
     context 'for method step' do
@@ -41,7 +46,8 @@ RSpec.describe ServiceOperator::Step do
       before do
         allow(service_class).to receive(:new).and_return(service_object)
         allow(service_object).to receive(:public_send)
-        allow(service_object).to receive(:method).with(call_parameters_method_name).and_return(parameters_list_method)
+        allow(service_object).to receive(:public_send).with(call_parameters_method_name).and_return(parameters_list)
+        allow(service_object).to receive(:method).with(call_method_name).and_return(parameters_list_method)
         allow(parameters_list_method).to receive(:parameters).and_return(parameters_list)
         allow(operator).to receive(:context).and_return(context)
         allow(context).to receive(:fail)
@@ -49,7 +55,7 @@ RSpec.describe ServiceOperator::Step do
 
       context 'for default configuration' do
         let(:call_method_name) { :call }
-        let(:call_parameters_method_name) { :call }
+        let(:call_parameters_method_name) { nil }
 
         context 'with empty call method parameters' do
           let(:parameters_list) { [] }
@@ -110,11 +116,9 @@ RSpec.describe ServiceOperator::Step do
         before do
           allow(service_object).to receive(:public_send).with(:failure?).and_return(call_result)
 
-          operator.configure do |config|
-            config.call_method_name = call_method_name
-            config.call_parameters_method_name = call_parameters_method_name
-            config.failure_method_name = failure_method_name
-          end
+          configuration.call_method_name = call_method_name
+          configuration.call_parameters_method_name = call_parameters_method_name
+          configuration.failure_method_name = failure_method_name
         end
 
         context 'without failure method' do
@@ -137,6 +141,7 @@ RSpec.describe ServiceOperator::Step do
             it 'calls service with parameters', :aggregate_failures do
               step_run
 
+              expect(service_object).to have_received(:public_send).with(:call_parameters)
               expect(service_object).to have_received(:public_send).with(:just_call, week_id: 1)
               expect(context).not_to have_received(:fail)
             end
@@ -163,6 +168,7 @@ RSpec.describe ServiceOperator::Step do
             it 'calls service with parameters', :aggregate_failures do
               step_run
 
+              expect(service_object).to have_received(:public_send).with(:call_parameters)
               expect(service_object).to have_received(:public_send).with(:just_call, week_id: 1)
               expect(context).not_to have_received(:fail)
             end